# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rocket Eventos",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/admin/events/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Deletes an event.",
                "operationId": "ffeb53815e6ed101fe7cc0bcd7fd8bef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the event.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The event was deleted."
                    },
                    "400": {
                        "description": "When \"code\" is zero, means validation. When is one, means that the event does not exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The admin is not authenticated."
                    },
                    "403": {
                        "description": "The user is not an admin/super."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Updates an event.",
                "operationId": "b1c734bd4f72fa8172746275f464a361",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the event.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": ""
                                    },
                                    "slug": {
                                        "type": "string",
                                        "default": ""
                                    },
                                    "details": {
                                        "type": "string",
                                        "default": ""
                                    },
                                    "subscription_date_start": {
                                        "type": "string",
                                        "format": "date-time",
                                        "default": ""
                                    },
                                    "subscription_date_end": {
                                        "type": "string",
                                        "format": "date-time",
                                        "default": ""
                                    },
                                    "presentation_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The event was updated and returns all its data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "When \"code\" is zero, means validation. When is one, means that the event does not exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The admin is not authenticated."
                    },
                    "403": {
                        "description": "The user is not an admin/super."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/events/{slug}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Details of the event.",
                "operationId": "7087984accf1312cf3c0a45572c69c39",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the event.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "slug"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The event was created and returns all its data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "When \"code\" is zero, means validation. When is one, means that the event does not exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The admin is not authenticated."
                    },
                    "403": {
                        "description": "The user is not an admin/super."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "EventModel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1000"
                    },
                    "name": {
                        "type": "string",
                        "example": "My Event"
                    },
                    "slug": {
                        "type": "string",
                        "example": "my-event"
                    },
                    "details": {
                        "type": "string",
                        "example": "A long text."
                    },
                    "subscriptionStart": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T20:13:42.529682Z"
                    },
                    "subscriptionEnd": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T20:13:42.529697Z"
                    },
                    "presentationAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T20:13:42.529706Z"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T20:13:42.529715Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T20:13:42.529723Z"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}